formatted main, moved functions to new saveLoad files
